feat(db-custom-types.ts): add custom type TsRange and custom type tsrange to handle PostgreSQL tsrange data type
The custom type TsRange is added to handle a range of dates in the form of a start and end date. It provides getter methods for the start and end values of the range.

The custom type tsrange is added as a customType in drizzle-orm/pg-core to handle the PostgreSQL tsrange data type. It provides conversion methods from the driver value to the custom type TsRange and vice versa. The fromDriver method parses the string value from the driver and creates a new TsRange object. The toDriver method serializes the TsRange object to a string value that can be sent to the driver.