feat: add contact form functionality to send emails
The contact form component was added to allow users to send emails. The form includes fields for name, email, subject, and message. When the form is submitted, the input values are sent to the server using the `sendMailAction` function. This function uses the `nodemailer` library to send the email using the Gmail service. The email is sent to the admin email address specified in the `ADMIN_EMAIL_ADDRESS` constant. The email includes the contact's name, email, and message.