chore(tailwind): update background color shades and add new accent color
The background color shades in the tailwind base configuration have been updated. The new shade "background-200" has been added with a value of "hsl(var(--background-200))". This change improves the flexibility and consistency of the background color usage.

Additionally, a new accent color "accents-500" has been introduced with a value of "hsl(var(--accents-500))". This new accent color can be used to add visual interest and variety to the UI.

These changes enhance the overall design and customization options of the tailwind CSS framework.