feat(ui-descriptions): add support for vertical layout in Descriptions component
The Descriptions component now supports a new prop called "layout" which can be set to "horizontal" or "vertical". When set to "vertical", the component will render the labels and children of each item in separate rows, making it easier to read and understand the information. This layout is especially useful when there are a large number of items in the Descriptions component.