feat: add Collapse component for creating collapsible sections in UI
The Collapse component provides a way to create collapsible sections in the UI. It is built on top of the @radix-ui/react-accordion library. The component accepts the following props:- `activeKey`: An array of keys representing the currently active items.- `defaultActiveKey`: An array of keys representing the default active items.- `items`: An array of objects representing the collapsible items. Each item object should have a `label` property for the item label and a `children` property for the content of the item.- `type`: The type of collapse, either "single" or "multiple". In "single" mode, only one item can be expanded at a time. In "multiple" mode, multiple items can be expanded simultaneously.- `triggerProps`: Additional props to be passed to the AccordionTrigger component.- `contentProps`: Additional props to be passed to the Accordion