Temp (#29)
* fix: update Pagination component to use 'page' instead of 'current' for consistency and clarity

* refactor: update input component to support different sizes for input fields

* feat: add Breadcrumbs component with dynamic rendering of items and separators

* refactor: restructure Popover component to support different trigger types and content display

* refactor: update DatePicker component to use ahooks for click away functionality and improve date handling and formatting

* chore: update imports and type definitions for Select component and its subcomponents

This commit updates the import paths for some types and components used in the Select component. It also updates the type definitions for the Select component and its subcomponents to align with the changes in the import paths.

* refactor: remove unused trigger prop and use selectDefaultPlaceholder function for placeholder in AutoComplete component

* style: remove disabled styles from button variants and add disabled shape variant to button component for better styling consistency

* refactor: move components to separate file for better organization

* refactor: improve date handling in DatePickerInternal component

* fix: exclude "size" attribute from InputProps type to prevent conflicts with VariantProps

* feat(input): add ability to clear input content with clear icon and implement triggerNativeEventFor function

The changes introduce the ability to clear input content with a clear icon by adding the `allowClear` prop to the `Input` component. Additionally, a new `triggerNativeEventFor` function has been implemented to simulate native events on elements, inspired by a Stack Overflow post.

* fix(date-picker): refactor date picker components and update date range picker functionality

* refactor(ui/table): streamline column definitions and improve table component structure

The changes in this commit involve refactoring the column definitions in the
`transformColumnDefs` function to simplify the logic and improve readability.
Additionally, the `TableProps` type has been updated to include a `rowClassName`
property for custom row styling. The `TableInner` component has been enhanced
to handle custom row rendering and styling based on the provided `rowClassName`.
These changes aim to improve the maintainability and flexibility of the table
component in the application.

* style(ui-calendar): update styling for selected day in Calendar component

* feat(search): add debouncing to search input to improve performance and user experience

- Add debounce functionality to handleSearch function in Search component to
  delay search execution and improve performance.
- Update dependencies to include 'ahooks' library for debouncing functionality.

* refactor(pagination): update PaginationLink component to use Next.js Link component for better navigation performance

* refactor(pagination): update Pagination component to use next/navigation hooks and remove unused code

* refactor(table): refactor TableInner component to use React.useMemo for columns and data to improve performance

* style(table): support w-full and fixed column width
improve Table component layout and add column sizing functionality

* style(table-head): remove unnecessary styles and add new styles for better readability and aesthetics

* feat(table): fixed columns

* feat(table-sticky): update Table component to support configurable sticky header properties

* refactor(checkbox): update Checkbox component to support indeterminate state and improve accessibility and styling

* style(tailwind): update color variables for ds-gray to improve consistency and readability

* feat(table): add row selection functionality to Table component

The changes introduce the ability to add row selection to the Table component.
The `transformColumnDefs` function now accepts a `rowSelection` prop, which, when provided, adds a selection column to the table. The `createSelectColumn` function creates the selection column with checkboxes for selecting rows. The `getRowRange` function helps in selecting a range of rows when using shift + click. Additionally, the `TableProps` type now includes a `rowSelection` property for configuring row selection.

* refactor(tooltip): update Tooltip component styling and behavior for better user experience

* feat(table): Add sorting functionality to the table component

The changes introduce a new `sorter` property to the `Meta` type, allowing for
local sorting with various built-in sorting functions or custom sorting
functions. Additionally, the code now includes the necessary logic to handle
sorting changes in the table component, allowing users to sort the table data
based on different columns.

* feat(table): add localization support for Vietnamese and English languages in Table component

* fix(table): add missing import for tableLocale_vi and tableLocale_en in localization files

* feat(ui): Update AutoComplete and Command components

- Added ValueType import to AutoComplete component
- Updated AutoCompleteProps and AutoCompleteInner to use ValueType
- Updated CommandProps and Command to use ValueType
- Updated CommandItem key and value props to use string for consistency
- Improved usePagination hook to handle page and pageSize changes
- Refactored ProjectDetailPage component to use Suspense and useState