feat(tailwind): Add new background color variables
The changes add two new CSS variables for the Tailwind design system
background colors, `--ds-background-100` and `--ds-background-200`. These
variables are then used in the `baseColors` object in the `base.ts` file to
define the `surface` and `surface.secondary` colors.

Additionally, the `--primary-hover` variable has been renamed to
`--primary-700` to better reflect the color it represents.