feat(ui/checkbox): update checkbox styles
Updates the checkbox styles to use a more modern design. The border
color has been updated to use a gray-700 color, and a transition for
the colors has been added. The hover style has also been updated to use
a background-hover color.

These changes were made to improve the overall visual appearance of the
checkbox component and align it with the design system.